+++ packages/fluo/archive/PresetCollection.js
@@ -1,10 +1,10 @@
-import { FRESH, PLANET, SUBTLE }                                from '@palett/presets'
-import { stringValue }                                          from '@texting/string-value'
+import { FRESH, PLANET, SUBTLE } from '@palett/presets'
+import { stringValue }           from '@texting/string-value'
 import { isNumeric as isNumericFull, parseNum as parseNumFull } from '@texting/charset-fullwidth'
 import { isNumeric, parseNum }                                  from '@texting/charset-halfwidth'
 import { isLiteral, isLiteralAny }                              from '@typen/literal'
 import { replenish }                                            from '@vect/object-update'
-import { mutate, mutazip }                                      from '@vect/vector'
+import { mutate, mutazip }                                      from '@vect/vector-mapper'
 
 export const isNumericAny = x => isNumericFull(x) || isNumeric(x)