+++ packages/munsell/test/RGB.unit.js
@@ -3,7 +3,7 @@ import { HexDye }             from '@palett/dye'
 import { Fluo }               from '@palett/fluo'
 import { hexToStr, rgbToStr } from '@palett/stringify'
 import { DecoObject, says } from '@spare/logger'
-import { RGB }              from '../src/extends/RGB.js'
+import { RGB }              from '../archive/extends/RGB.js'
 
 export const XTERM = {
   noir: [ 0, 0, 0 ],