+++ packages/pres/index.js
@@ -1,12 +1,10 @@
-export { Pres }           from './src/Pres.js'
-export { Presm }          from './src/Presm.js'
-export { demo, randPres } from './src/pres-extensions.js'
-
-export { Pres as Preset }         from './src/Pres.js'
-export { randPres as randPreset } from './src/pres-extensions.js'
+export { Pres, Pres as Preset } from './src/Pres.js'
+export { Presm }                from './src/Presm.js'
+export { demo, sequence }       from './src/sequence.js'
+export { randPres }             from './src/randPres.js'
 
 export {
-  hexToHsi, rgaToHsi, hsaToRgi, hsiToHsl, hsiToRgi, modHsi, modHsiTo, modRgi, deltaHsi, hsaToInt, hsiToInt
+  hexToHsi, rgaToHsi, hsaToRgi, hsiToHsl, hsiToRgi, modHsi, modHsiTo, modRgi, deltaHsi, hsaToInt, hsiToInt,
 } from '@palett/convert'