+++ packages/pres/test/presets.test.js
@@ -5,7 +5,7 @@ import {
 import { logger }  from '@spare/logger'
 import { indexed } from '@vect/object-mapper'
 import { test } from 'node:test'
-import { demo } from '../src/pres-extensions.js'
+import { demo } from '../src/sequence.js'
 
 const Pavtone = {
   RODD: RODD,