+++ constants/enum-colorant-modes/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @palett/enum-colorant-modes
 
+## 0.12.14
+
+### Patch Changes
+
+- minor fix
+- Updated dependencies
+  - @palett/enum-color-space@0.12.14
+
 ## 0.12.13
 
 ### Patch Changes