+++ constants/enum-font-effects/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @palett/enum-font-effects
 
+## 0.12.14
+
+### Patch Changes
+
+- minor fix
+
 ## 0.12.13
 
 ### Patch Changes