+++ packages/projector/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @palett/projector
 
+## 0.12.14
+
+### Patch Changes
+
+- minor fix
+- Updated dependencies
+  - @palett/dye@0.12.14
+  - @palett/presets@0.12.14
+
 ## 0.12.13
 
 ### Patch Changes